+++ game/core/src/ui/LeftSideBar/index.tsx
@@ -2,19 +2,12 @@ import { TeamType } from "@heroesofcrypto/common";
 import DiceIcon from "@mui/icons-material/Casino";
 import DashboardRoundedIcon from "@mui/icons-material/DashboardRounded";
 import FactoryRoundedIcon from "@mui/icons-material/FactoryRounded";
-import InfoRoundedIcon from "@mui/icons-material/InfoRounded";
 import OpenInNewRoundedIcon from "@mui/icons-material/OpenInNewRounded";
-import RefreshRoundedIcon from "@mui/icons-material/RefreshRounded";
 import TerrainRoundedIcon from "@mui/icons-material/TerrainRounded";
-import TimelapseRoundedIcon from "@mui/icons-material/TimelapseRounded";
-import ZoomInMapIcon from "@mui/icons-material/ZoomInMap";
 import Avatar from "@mui/joy/Avatar";
 import Badge from "@mui/joy/Badge";
 import Box from "@mui/joy/Box";
-import Button from "@mui/joy/Button";
-import Card from "@mui/joy/Card";
 import Divider from "@mui/joy/Divider";
-import LinearProgress from "@mui/joy/LinearProgress";
 import List from "@mui/joy/List";
 import ListItem from "@mui/joy/ListItem";
 import ListItemButton from "@mui/joy/ListItemButton";
@@ -28,8 +21,9 @@ import * as packageJson from "../../../package.json";
 import { images } from "../../generated/image_imports";
 import { useManager } from "../../manager";
 import { IVisibleState, IVisibleUnit } from "../../state/visible_state";
-import { BAR_SIZE_PIXELS_STR } from "../../statics";
+import { EDGES_SIZE } from "../../statics";
 import ColorSchemeToggle from "./ColorSchemeToggle";
+import { MessageBox } from "./MessageBox";
 import { UnitStatsListItem } from "./UnitStatsListItem";
 
 const UpNext: React.FC = () => {
@@ -117,135 +111,9 @@ const UpNext: React.FC = () => {
     );
 };
 
-const MessageBox = ({ gameStarted }: { gameStarted: boolean }) => {
-    const [visibleState, setVisibleState] = useState<IVisibleState>({} as IVisibleState);
-
-    const manager = useManager();
-
-    useEffect(() => {
-        const connection = manager.onVisibleStateUpdated.connect(setVisibleState);
-        return () => {
-            connection.disconnect();
-        };
-    }, [manager]);
-
-    let messageBoxVariant: "plain" | "outlined" | "soft" | "solid" | undefined;
-    let messageBoxColor: "primary" | "neutral" | "danger" | "success" | "warning" | undefined;
-    let messageBoxTitle;
-    let messageBoxText;
-    let messageBoxButtonText;
-    let messageBoxRequestAdditionalTimeButtonRendered = false;
-    let messageBoxProgressBar: React.JSX.Element;
-    let messageBoxProgressValue = gameStarted ? 0 : 80;
-    if (visibleState.secondsMax) {
-        messageBoxProgressValue = 100 - (visibleState.secondsRemaining / visibleState.secondsMax) * 100;
-    }
-    const progress = (
-        <LinearProgress
-            variant="outlined"
-            determinate={gameStarted}
-            value={messageBoxProgressValue}
-            sx={{ my: 1, overflow: "hidden" }}
-        />
-    );
-    const defaultIcon =
-        visibleState.lapNumber &&
-        visibleState.lapNumber < visibleState.numberOfLapsTillStopNarrowing &&
-        !(visibleState.lapNumber % visibleState.numberOfLapsTillNarrowing) ? (
-            <ZoomInMapIcon />
-        ) : (
-            <TimelapseRoundedIcon />
-        );
-
-    if (gameStarted) {
-        messageBoxVariant = "soft";
-        if (visibleState.hasFinished) {
-            messageBoxColor = "neutral";
-            messageBoxTitle = "Fight finished";
-            messageBoxText = "Refresh the page to start a new one";
-            messageBoxButtonText = "";
-            messageBoxProgressBar = <span />;
-        } else {
-            if (messageBoxProgressValue <= 45) {
-                messageBoxColor = "success";
-                messageBoxButtonText = "";
-            } else if (messageBoxProgressValue <= 70) {
-                messageBoxColor = "warning";
-                messageBoxButtonText = "";
-            } else {
-                messageBoxColor = "danger";
-                if (visibleState.canRequestAdditionalTime) {
-                    messageBoxButtonText = "Use additional time";
-                    messageBoxRequestAdditionalTimeButtonRendered = true;
-                }
-            }
-            messageBoxTitle = `Lap ${visibleState.lapNumber}`;
-            if (!visibleState.teamTypeTurn) {
-                messageBoxText = "Calculating next turn.";
-            } else if (visibleState.teamTypeTurn === TeamType.LOWER) {
-                messageBoxText = "Green team is making a turn";
-            } else {
-                messageBoxText = "Red team is making a turn";
-            }
-            messageBoxProgressBar = progress;
-        }
-    } else {
-        messageBoxProgressBar = progress;
-        messageBoxVariant = "solid";
-        messageBoxColor = "primary";
-        messageBoxTitle = "To start";
-        messageBoxText =
-            "Place the units from both teams on the board. Make sure to have at least one unit on both the top and bottom sides.";
-        if (visibleState.canBeStarted) {
-            messageBoxButtonText = "Start";
-        } else {
-            messageBoxButtonText = "";
-        }
-    }
-
-    return (
-        <Card invertedColors variant={messageBoxVariant} color={messageBoxColor} size="sm" sx={{ boxShadow: "none" }}>
-            <Stack direction="row" justifyContent="space-between" alignItems="center">
-                <Typography level="title-sm">{messageBoxTitle}</Typography>
-                {(() => {
-                    if (gameStarted) {
-                        if (visibleState.hasFinished) {
-                            return <RefreshRoundedIcon />;
-                        }
-                        return defaultIcon;
-                    }
-                    return <InfoRoundedIcon />;
-                })()}
-            </Stack>
-            <Typography level="body-xs">{messageBoxText}</Typography>
-            {messageBoxProgressBar}
-
-            {messageBoxButtonText ? (
-                <Button
-                    onClick={() =>
-                        messageBoxRequestAdditionalTimeButtonRendered
-                            ? manager.RequestTime(visibleState.teamTypeTurn)
-                            : manager.StartGame()
-                    }
-                    onMouseDown={() =>
-                        messageBoxRequestAdditionalTimeButtonRendered
-                            ? manager.RequestTime(visibleState.teamTypeTurn)
-                            : manager.StartGame()
-                    }
-                    size="sm"
-                    variant="solid"
-                >
-                    {messageBoxButtonText}
-                </Button>
-            ) : (
-                <span />
-            )}
-        </Card>
-    );
-};
-
 export default function LeftSideBar({ gameStarted }: { gameStarted: boolean }) {
     const [badgeVisible, setBadgeVisible] = useState(false);
+    const [barSize, setBarSize] = useState(280); // Initialize bar size state
     const [buttonsVisible] = useState({
         prediction: false,
         terrain: false,
@@ -254,6 +122,20 @@ export default function LeftSideBar({ gameStarted }: { gameStarted: boolean }) {
     });
     const { setMode } = useColorScheme();
     const theme = useTheme();
+
+    const adjustBarSize = () => {
+        const additionalBoardPixels = gameStarted ? 0 : 512;
+        const edgesSize = gameStarted ? 0 : EDGES_SIZE;
+        const widthRatio = window.innerWidth / (2048 + edgesSize + additionalBoardPixels);
+        const heightRatio = window.innerHeight / (2048 + edgesSize);
+
+        const scaleRatio = Math.min(widthRatio, heightRatio);
+        const scaledBoardSize = (2048 + additionalBoardPixels) * scaleRatio;
+
+        const edgeSizeWidth = gameStarted ? 0 : edgesSize / 2;
+        const rightBarEndAtBoard = (window.innerWidth - scaledBoardSize) / 2;
+        setBarSize(rightBarEndAtBoard > edgeSizeWidth ? rightBarEndAtBoard : edgeSizeWidth);
+    };
 
     useEffect(() => {
         const interval = setInterval(() => {
@@ -271,6 +153,14 @@ export default function LeftSideBar({ gameStarted }: { gameStarted: boolean }) {
         setMode("dark"); // Set dark mode by default
     }, [setMode]);
 
+    // Adjust bar size on window resize
+    useEffect(() => {
+        window.addEventListener("resize", adjustBarSize);
+        adjustBarSize(); // Initial call to set the size based on the initial window dimensions
+
+        return () => window.removeEventListener("resize", adjustBarSize);
+    }, [gameStarted]);
+
     // @ts-ignore: skip styles
     return (
         <Sheet
@@ -279,7 +169,7 @@ export default function LeftSideBar({ gameStarted }: { gameStarted: boolean }) {
                 position: "fixed",
                 zIndex: 1,
                 height: "100dvh",
-                width: BAR_SIZE_PIXELS_STR,
+                width: `${barSize}px`, // Use dynamic bar size
                 top: 0,
                 left: 0,
                 p: 2,
